drivers/bus/spi-bus: Add an SPI bus interface.
Every SPI peripherals are no longer in charge of managing the SPI
peripherals. They are juste writing to the SPI bus they are attached to.

A typical access to the SPI bus is as following:

    	/* Wait the bus to be free */
    	while (spi_bus_transaction_pending(bus_id));
    	/* Start a transaction on the SPI bus */
    	spi_bus_start_transaction(bus_id);

    	/* Transfer buffers as many as needed */
    	status = spi_bus_write(bus_id, chip_select, &out, 1, NULL, NULL);
    	if (status) {
    		twi_bus_stop_transaction(bus);
    		return status;
    	}
    	/* Wait for the tranisaction to finish */
    	spi_bus_wait_transfer(bus_id));
    	/* Release the SPI bus */
    	spi_bus_stop_transaction(bus_id);

+++ drivers/peripherals/spi.h
@@ -129,7 +129,7 @@ extern void spi_release_cs(Spi * spi);
  * \param delay_dlybct
  * \param spi_mode
  */
-extern void spi_configure_cs(Spi * spi, uint32_t cs, uint32_t bitrate,
+extern void spi_configure_cs(Spi * spi, uint8_t cs, uint32_t bitrate,
 			     uint32_t delay_dlybs, uint32_t delay_dlybct,
 			     uint32_t spi_mode);
